requirements.txt
# Core dependencies for the AI research agent
atomic-agents>=1.0.24
instructor>=1.3.7
google-generativeai>=0.8.0
python-dotenv>=1.0.1

# Web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation and HTTP client
pydantic>=2.0.0
httpx>=0.25.0

# Development and testing dependencies
pytest>=8.3.5
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Code quality and linting
mypy>=1.11.1
ruff>=0.6.1

# Build system
setuptools>=73.0.0
wheel